2010-04-20	Ralf Corsépius <ralf.corsepius@rtems.org>
	* console/console.c: Reflect termios-API having changed.

+++ c/src/lib/libbsp/sh/gensh4/console/console.c
@@ -65,8 +65,8 @@ console_poll_read(int minor)
  * RETURNS:
  *     result code
  */
-static int
-console_interrupt_write(int minor, const char *buf, int len)
+static ssize_t
+console_interrupt_write(int minor, const char *buf, size_t len)
 {
         return sh4uart_interrupt_write(&sh4_uarts[minor], buf, len);
 }
@@ -82,8 +82,8 @@ console_interrupt_write(int minor, const char *buf, int len)
  * RETURNS:
  *     result code
  */
-static int
-console_poll_write(int minor, const char *buf, int len)
+static ssize_t
+console_poll_write(int minor, const char *buf, size_t len)
 {
         return sh4uart_poll_write(&sh4_uarts[minor], buf, len);
 }
